fix: [#54992] fixing os2syncs imports
There is an error where os2sync tries to import lora cache,
which it can't since the new caches dependencies are not
installed.

This fix imports everything that the sql export does, and what
os2sync used to. This should get os2sync up and running, and then
we can clean up the installed dependencies when Benjamin or Kari
is back at work.