feat: [#62639] add endpoint to passivate and re-sync all objects
In some cases there can be duplicates on addresses and engagements
in fk-org. Theese can only be removed by deactivating the object and
syncing it again.

This commit adds and endpoint which runs through every orgunit and user
in MO and calls os2syncs 'passiver' endpoint and then syncs the data again.